Fix inconsistent common variable return ordering
