+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.black]
 line-length = 79
-extend-exclude = '(__snapshots__|integration/slices|integration/sources|integration/envs|outputs|_alembic)'
+extend-exclude = '(__snapshots__|integration/slices|integration/sources|integration/envs|outputs|_alembic|unit/plugins/expected)'
 
 [tool.isort]
 profile = "black"
