+++ tests/unit/plugins/expected/airflow_complex_h_perart/airflow_complex_h_perart_Dockerfile
